requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
distro==1.9.0
exceptiongroup==1.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==5.4.0
openai==1.82.0
pillow==11.2.1
pydantic==2.11.5
pydantic_core==2.33.2
python-dotenv==1.1.0
python-pptx==1.0.2
requests==2.32.3
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
XlsxWriter==3.2.3